backend/nlp/requirements.txt
fastapi
uvicorn
requests
groq
python-dotenv
pydantic
